feat: Refactor deployment URLs in DeploymentCard.svelte and +layout.svelte files
