+++ frontend/src/routes/app/blueprints/+layout.svelte
@@ -16,12 +16,12 @@
     in:slide={{ duration: 200, easing: cubicOut }}
     out:slide={{ duration: 200, easing: cubicOut }}
   >
-    <div class="absolute top-10 left-2 bottom-10 overflow-y-hidden">
+    <div class="absolute top-5 left-2 bottom-10 overflow-y-hidden">
       <BlueprintSideNav />
     </div>
     {#key data.url}
       <div
-        class="absolute top-4 pt-4 left-64 right-2 bottom-0 overflow-y-auto scrollbar-none pb-8 px-2 overflow-x-hidden"
+        class="absolute top-0 pt-4 left-64 right-2 bottom-0 overflow-y-auto scrollbar-none pb-8 px-2 overflow-x-hidden"
         in:fly={{ duration: 100, easing: cubicOut }}
         out:fly={{ duration: 100, easing: cubicOut }}
       >
