+++ frontend/src/routes/app/projects/[id]/+page.svelte
@@ -1,15 +1,29 @@
 <script lang="ts">
+  import { goto } from "$app/navigation";
   import DeploymentCard from "$lib/components/deployments/DeploymentCard.svelte";
   import { deployments } from "$lib/stores/data";
-  import { Heading } from "flowbite-svelte";
+  import { Button, Heading } from "flowbite-svelte";
+  import { BookDashed } from "lucide-svelte";
 </script>
 
 <div class="absolute w-full top-0 bottom-0 overflow-y-scroll scrollbar-none">
   <div class="max-w-6xl mx-auto px-5 pb-5 py-5 flex flex-col">
     <div class="flex mb-5">
       <Heading tag="h5" class="flex font-normal items-center w-auto"
-        >Your Deployments ({$deployments.length})</Heading
+        >Deployments ({$deployments.length})</Heading
       >
+      <div class="justify-self-end ml-auto space-x-3">
+        <Button
+          color="primary"
+          size="sm"
+          on:click={() => {
+            goto("/app/blueprints/my-blueprints");
+          }}
+        >
+          <BookDashed class="w-4 h-4 mr-2 inline-block" />
+          Blueprints
+        </Button>
+      </div>
     </div>
 
     <ul role="list" class="divide-y dark:divide-white/5 divide-gray/5">
