+++ frontend/src/routes/app/projects/[id]/deployments/+page.svelte
@@ -1,4 +1,2 @@
 <script lang="ts">
 </script>
-
-lel
