+++ frontend/src/routes/app/projects/[id]/deployments/[id]/+layout.svelte
@@ -1,7 +1,9 @@
 <script lang="ts">
-  import { updateCurrentRolloutStatus } from "$lib/stores/data";
+  import { deployments, updateCurrentRolloutStatus } from "$lib/stores/data";
   import { onDestroy, onMount } from "svelte";
   import { navigating } from "$app/stores";
+  import { TabItem, Tabs } from "flowbite-svelte";
+  import { recordLogoUrl } from "$lib/utils/blueprint.utils";
 
   const updateCurrentRollout = () => {
     updateCurrentRolloutStatus();
@@ -25,5 +27,20 @@
     clearInterval(intervalId);
   });
 </script>
+
+<Tabs tabStyle="underline">
+  {#each $deployments as deployment}
+    <TabItem>
+      <div slot="title" class="flex items-center gap-2">
+        <img
+          src={recordLogoUrl(deployment)}
+          alt="Tuple"
+          class="h-9 w-9 flex-none rounded-lg object-cover p-1"
+        />
+        {deployment.name}
+      </div>
+    </TabItem>
+  {/each}
+</Tabs>
 
 <slot />
