+++ frontend/src/routes/app/projects_old/+layout.svelte
@@ -7,12 +7,12 @@
 
 {#if $page.url.pathname.startsWith("/app/projects/")}
   <div class="relative h-full max-w-6xl mx-auto">
-    <div class="absolute top-10 left-2 bottom-10 overflow-y-hidden">
+    <div class="absolute top-5 left-2 bottom-10 overflow-y-hidden">
       <SideNav />
     </div>
     {#key data.url}
       <div
-        class="absolute top-4 pt-4 left-64 right-2 bottom-0 overflow-y-auto scrollbar-none pb-8 px-2 overflow-x-hidden"
+        class="absolute top-0 pt-4 left-64 right-2 bottom-0 overflow-y-auto scrollbar-none pb-8 px-2 overflow-x-hidden"
       >
         <slot />
       </div>
